--- adam/Computations/__init__.py
@@ -1,1 +1,0 @@
-import adam.Computations.KinDynComputations
