--- adam/Geometry/__init__.py
@@ -1,1 +1,0 @@
-import adam.Geometry.utils
