+++ adam/__init__.py
@@ -1,2 +1,2 @@
-from adam.Computations import KinDynComputations
-from adam.Geometry import utils
+from . import core
+from . import geometry
