+++ adam/geometry/utils.py
