+++ setup.cfg
@@ -1,0 +1,21 @@
+[metadata]
+name = ADAM
+description = Automatic Differentiation for rigid body AlgorithMs
+author = "Giuseppe L'Erario"
+author_email = gl.giuseppelerario@gmail.com
+
+[options]
+python_requires = >=3.7
+install_requires =
+        numpy
+        scipy
+        casadi
+        matplotlib
+        prettytable
+        urdf_parser_py
+
+
+[options.extras_require]
+testing =
+    pytest
+    idyntree
